+++ ui/PluginWindow.hpp
@@ -9,7 +9,7 @@ namespace Musec
 {
 namespace UI
 {
-void createNewPluginWindow(std::shared_ptr<Musec::Audio::Plugin::IPlugin<double>> pluginNeedsWindow);
+void createNewPluginWindow(std::shared_ptr<Musec::Audio::Plugin::IPlugin<float>> pluginNeedsWindow);
 void createNewPluginWindowComplete(QWindow* window);
 }
 }
